requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-secure==1.0.2
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
dj-database-url
dj-static
python-decouple
gunicorn
psycopg2-binary
whitenoise